add ProductNotFoundException

+++ Services/Catalog/Catalog.Application/Handlers/DeleteProductByIdHandler.cs
@@ -1,3 +1,4 @@
+using Catalog.Application.Exceptions;
 using Catalog.Application.Queries;
 using Catalog.Core.Repositories;
 using MediatR;
@@ -15,6 +16,11 @@ public class DeleteProductByIdHandler : IRequestHandler<DeleteProductByIdQuery, 
 
     public async Task<bool> Handle(DeleteProductByIdQuery request, CancellationToken cancellationToken)
     {
+        var productToDelete = await _productRepository.GetProduct(request.Id);
+
+        if (productToDelete == null)
+            throw new ProductNotFoundException(request.Id);
+
         return await _productRepository.DeleteProduct(request.Id);
     }
 }
